requirements.txt
uvicorn==0.35.0
sqlalchemy==2.0.43
fastapi==0.116.1
PyJWT==2.10.1
cryptography==45.0.7
jinja2==3.1.6
pydantic-settings==2.12.0